#507: Denied deletion of milestones for those associated with observations and mixed the order index starting number from 1 to 0

+++ studymanager-services/src/main/java/io/redlink/more/studymanager/service/MilestoneService.java
@@ -57,7 +57,7 @@ public class MilestoneService {
 
     private void reorder(long studyId, int milestoneId, int oldIndex, int requestedNewIndex) {
         int count = repository.countByStudyId(studyId);
-        int newIndex = Math.max(1, Math.min(requestedNewIndex, count));
+        int newIndex = Math.max(0, Math.min(requestedNewIndex, count - 1));
         if (newIndex == oldIndex) {
             return;
         }
@@ -77,6 +77,9 @@ public class MilestoneService {
         Milestone milestone = getMilestone(studyId, milestoneId);
         if (repository.countActiveParticipantMilestones(studyId, milestoneId) > 0) {
             throw DataConstraintException.createMilestoneInUseByActiveParticipant(studyId, milestoneId);
+        }
+        if (repository.countObservationsUsingMilestone(studyId, milestoneId) > 0) {
+            throw DataConstraintException.createMilestoneInUseByObservation(studyId, milestoneId);
         }
         repository.deleteById(studyId, milestoneId);
         repository.decrementOrderIndexAbove(studyId, milestone.getOrderIndex());
